Add get instance and list instances
